update custom meta path finder
